Issue 78: Move the RolesManager as an application query
