+++ src/Khatovar/Bundle/UserBundle/Resources/config/controllers.yml
@@ -33,7 +33,7 @@ services:
             - '@Khatovar\Component\User\Application\Query\GetUserFromDatabase'
             - '@khatovar_user.factory.user_form'
             - '@khatovar_user.manager.users'
-            - '@khatovar_user.manager.roles'
+            - '@Khatovar\Component\User\Application\Query\UserRole'
             - '@security.token_storage'
             - '@event_dispatcher'
             - '@session'
