+++ src/Khatovar/Bundle/UserBundle/Resources/config/queries.yml
@@ -3,3 +3,8 @@ services:
         arguments:
             - '@Khatovar\Bundle\UserBundle\Entity\Repository\UserRepository'
             - '@Khatovar\Bundle\UserBundle\Security\Authentication\UserInSecurityTokenStorage'
+
+    Khatovar\Component\User\Application\Query\UserRole:
+        arguments:
+            - '@Khatovar\Bundle\UserBundle\Security\Authentication\UserInSecurityTokenStorage'
+            - '%security.role_hierarchy.roles%'
