+++ tests/spec/Khatovar/Component/User/Application/Query/UserRoleSpec.php
@@ -9,17 +9,17 @@
  * file that was distributed with this source code.
  */
 
-namespace spec\Khatovar\Bundle\UserBundle\Manager;
+namespace spec\Khatovar\Component\User\Application\Query;
 
 use Khatovar\Component\User\Domain\Model\UserInterface;
-use Khatovar\Bundle\UserBundle\Manager\RolesManager;
+use Khatovar\Component\User\Application\Query\UserRole;
 use Khatovar\Component\User\Application\Query\CurrentTokenUser;
 use PhpSpec\ObjectBehavior;
 
 /**
  * @author Damien Carcel <damien.carcel@gmail.com>
  */
-class RolesManagerSpec extends ObjectBehavior
+class UserRoleSpec extends ObjectBehavior
 {
     function let(CurrentTokenUser $currentTokenUser)
     {
@@ -35,14 +35,14 @@ class RolesManagerSpec extends ObjectBehavior
 
     function it_is_a_roles_manager()
     {
-        $this->shouldHaveType(RolesManager::class);
+        $this->shouldHaveType(UserRole::class);
     }
 
     function it_returns_the_list_of_roles_for_the_super_admin($currentTokenUser)
     {
         $currentTokenUser->isSuperAdmin()->willReturn(true);
 
-        $this->getChoices()->shouldReturn([
+        $this->listAvailableOnes()->shouldReturn([
             'ROLE_USER'   => 'ROLE_USER',
             'ROLE_VIEWER' => 'ROLE_VIEWER',
             'ROLE_ADMIN'  => 'ROLE_ADMIN',
@@ -53,7 +53,7 @@ class RolesManagerSpec extends ObjectBehavior
     {
         $currentTokenUser->isSuperAdmin()->willReturn(false);
 
-        $this->getChoices()->shouldReturn([
+        $this->listAvailableOnes()->shouldReturn([
             'ROLE_USER'   => 'ROLE_USER',
             'ROLE_VIEWER' => 'ROLE_VIEWER',
         ]);
